backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyMySQL==1.1.0
Werkzeug==3.0.1
PyJWT==2.8.0
async-playwright==1.39.0
aiohttp==3.9.1
celery==5.3.6
redis==5.0.1
eventlet==0.33.3
